Version finale de dev

+++ lib/screens/list_view/list_view.dart
@@ -1,6 +1,10 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
+import 'package:todolist/classes/todo.dart';
+import 'package:todolist/classes/task.dart';
+import 'package:todolist/components/list_form.dart';
+import 'package:flutter_tags/flutter_tags.dart';
 import 'package:intl/intl.dart';
 
 class ListViewScreen extends StatefulWidget {
@@ -12,13 +16,15 @@ class _ListViewScreenState extends State<ListViewScreen> {
   final Future<FirebaseApp> _initialization = Firebase.initializeApp();
   final _formKey = GlobalKey<FormState>();
   Task _task = new Task(null);
+  Todo todo = new Todo();
 
-  Widget _buildName() {
+  Widget _taskForm(currentValue) {
     return TextFormField(
+      initialValue: currentValue != null ? currentValue : '',
       decoration: InputDecoration(labelText: "Nom de la tâche"),
       validator: (String value) {
         if (value.isEmpty) {
-          return "required";
+          return "Requis";
         }
         return null;
       },
@@ -30,12 +36,16 @@ class _ListViewScreenState extends State<ListViewScreen> {
 
   @override
   Widget build(BuildContext context) {
-    final String args = ModalRoute.of(context).settings.arguments;
+    this.todo = ModalRoute.of(context).settings.arguments;
 
     CollectionReference taches = FirebaseFirestore.instance
         .collection('Listes')
-        .doc(args)
+        .doc(this.todo.documentId)
         .collection("Taches");
+
+    this.todo.documentReference = FirebaseFirestore.instance
+        .collection('Listes')
+        .doc(this.todo.documentId);
 
     CollectionReference listes =
         FirebaseFirestore.instance.collection('Listes');
@@ -48,11 +58,94 @@ class _ListViewScreenState extends State<ListViewScreen> {
           .catchError((error) => print("Failed to add Task: $error"));
     }
 
+    Future<void> deleteTask(taskId) {
+      // Call the Taches CollectionReference to add a new tache
+      return taches.doc(taskId).delete();
+    }
+
+    Future<bool> editTaskDialog(currentValue) async {
+      return showDialog<bool>(
+          context: context,
+          builder: (BuildContext context) {
+            return AlertDialog(
+              content: Stack(
+                overflow: Overflow.visible,
+                children: <Widget>[
+                  Positioned(
+                    right: -40.0,
+                    top: -40.0,
+                    child: InkResponse(
+                      onTap: () {
+                        Navigator.of(context).pop(false);
+                      },
+                      child: CircleAvatar(
+                        child: Icon(Icons.close),
+                        backgroundColor: Colors.red,
+                      ),
+                    ),
+                  ),
+                  Form(
+                    key: _formKey,
+                    child: Column(
+                      mainAxisSize: MainAxisSize.min,
+                      children: <Widget>[
+                        _taskForm(currentValue),
+                        RaisedButton(
+                          child: Text("Modifier"),
+                          onPressed: () {
+                            if (_formKey.currentState.validate()) {
+                              _formKey.currentState.save();
+                              Navigator.of(context).pop(true);
+                            }
+                          },
+                        ),
+                      ],
+                    ),
+                  ),
+                ],
+              ),
+            );
+          });
+    }
+
+    Future<bool> confirmDelete() async {
+      return showDialog<bool>(
+        context: context,
+        barrierDismissible: false, // user must tap button!
+        builder: (BuildContext context) {
+          return AlertDialog(
+            title: Text('Supprimer la liste'),
+            content: SingleChildScrollView(
+              child: ListBody(
+                children: <Widget>[
+                  Text('Êtes-vous sûr-e de vouloir supprimer la liste ?'),
+                ],
+              ),
+            ),
+            actions: <Widget>[
+              TextButton(
+                child: Text('Annuler'),
+                onPressed: () {
+                  Navigator.of(context).pop(false);
+                },
+              ),
+              TextButton(
+                child: Text('Supprimer'),
+                onPressed: () {
+                  Navigator.of(context).pop(true);
+                },
+              ),
+            ],
+          );
+        },
+      );
+    }
+
     taches.snapshots(includeMetadataChanges: true);
     return Scaffold(
       appBar: AppBar(
           title: FutureBuilder(
-              future: listes.doc(args).get(),
+              future: listes.doc(this.todo.documentId).get(),
               builder: (BuildContext context,
                   AsyncSnapshot<DocumentSnapshot> snapshot) {
                 // Check for errors
@@ -69,7 +162,7 @@ class _ListViewScreenState extends State<ListViewScreen> {
               }),
           actions: [
             FutureBuilder(
-                future: listes.doc(args).get(),
+                future: listes.doc(this.todo.documentId).get(),
                 builder: (BuildContext context,
                     AsyncSnapshot<DocumentSnapshot> snapshot) {
                   // Check for errors
@@ -78,12 +171,44 @@ class _ListViewScreenState extends State<ListViewScreen> {
                   }
 
                   if (snapshot.connectionState == ConnectionState.done) {
-                    Map<String, dynamic> data = snapshot.data.data();
+                    return IconButton(
+                      icon: Icon(Icons.edit),
+                      onPressed: () {
+                        return showDialog<Todo>(
+                            context: context,
+                            builder: (BuildContext context) {
+                              return ListAddForm(todo: this.todo);
+                            }).then((newValues) => {
+                              this.setState(() {
+                                this.todo = newValues;
+                              })
+                            });
+                      },
+                    );
+                  }
+
+                  return Text('Please wait');
+                }),
+            FutureBuilder(
+                future: listes.doc(this.todo.documentId).get(),
+                builder: (BuildContext context,
+                    AsyncSnapshot<DocumentSnapshot> snapshot) {
+                  // Check for errors
+                  if (snapshot.hasError) {
+                    return Text('Error');
+                  }
+
+                  if (snapshot.connectionState == ConnectionState.done) {
                     return IconButton(
                       icon: Icon(Icons.delete),
                       onPressed: () {
-                        listes.doc(snapshot.data.id).delete();
-                        Navigator.pop(context);
+                        confirmDelete().then((confirmed) => {
+                              if (confirmed)
+                                {
+                                  listes.doc(snapshot.data.id).delete(),
+                                  Navigator.pop(context)
+                                }
+                            });
                       },
                     );
                   }
@@ -103,22 +228,74 @@ class _ListViewScreenState extends State<ListViewScreen> {
               return Text("Loading");
             }
 
-            return new ListView(
-              children: snapshot.data.docs.map((DocumentSnapshot document) {
-                return new CheckboxListTile(
-                  title: Text(document.data()['nom']),
-                  controlAffinity: ListTileControlAffinity.leading,
-                  value: document.data()['fait'],
-                  onChanged: (bool value) {
-                    setState(() {
-                      taches.doc(document.id).update({'fait': value});
-                    });
-                  },
-                  activeColor: Colors.green,
-                );
-              }).toList(),
-            );
+            return ListView(
+                children: snapshot.data.docs.map((DocumentSnapshot document) {
+              return Row(children: [
+                Expanded(
+                    child: GestureDetector(
+                        onLongPress: () {
+                          editTaskDialog(document.data()['nom'])
+                              .then((confirmed) => {
+                                    if (confirmed)
+                                      {
+                                        setState(() {
+                                          taches
+                                              .doc(document.id)
+                                              .update({'nom': _task.nom});
+                                        })
+                                      }
+                                  });
+                        },
+                        child: CheckboxListTile(
+                          title: Text(document.data()['nom']),
+                          controlAffinity: ListTileControlAffinity.leading,
+                          value: document.data()['fait'],
+                          onChanged: (bool value) {
+                            setState(() {
+                              taches.doc(document.id).update({'fait': value});
+                            });
+                          },
+                          activeColor: Colors.green,
+                        ))),
+                IconButton(
+                    icon: Icon(Icons.delete),
+                    onPressed: () {
+                      deleteTask(document.id);
+                    })
+              ]);
+            }).toList());
           }),
+      bottomNavigationBar: BottomAppBar(
+        shape: const CircularNotchedRectangle(),
+        child: Container(
+            height: 90.0,
+            child: Padding(
+                padding: const EdgeInsets.all(10.0),
+                child: Column(children: [
+                  Row(children: [
+                    Text('Deadline: ' +
+                        DateFormat('dd MMMM yyyy à kk:mm')
+                            .format(this.todo.dLine) +
+                        '\n')
+                  ]),
+                  Row(children: [
+                    Tags(
+                      itemCount: this.todo.tags.length,
+                      itemBuilder: (int index) {
+                        return Padding(
+                            padding: const EdgeInsets.only(bottom: 5.0),
+                            child: ItemTags(
+                              key: Key(index.toString()),
+                              index: index,
+                              title: this.todo.tags[index],
+                              pressEnabled: false,
+                            ));
+                      },
+                    )
+                  ])
+                ]))),
+      ),
+      floatingActionButtonLocation: FloatingActionButtonLocation.endDocked,
       floatingActionButton: FloatingActionButton(
         onPressed: () {
           showDialog(
@@ -146,19 +323,16 @@ class _ListViewScreenState extends State<ListViewScreen> {
                         child: Column(
                           mainAxisSize: MainAxisSize.min,
                           children: <Widget>[
-                            _buildName(),
+                            _taskForm(null),
                             RaisedButton(
                               child: Text("Ajouter"),
                               onPressed: () {
                                 if (_formKey.currentState.validate()) {
                                   _formKey.currentState.save();
-                                  print(_task.nom);
-                                  print(args);
                                   addTask();
                                   Navigator.pop(context);
                                   return;
                                 }
-                                //print(_task.deadLine);
                               },
                             ),
                           ],
@@ -175,15 +349,5 @@ class _ListViewScreenState extends State<ListViewScreen> {
         ),
       ),
     );
-  }
-}
-
-class Task {
-  String nom;
-  bool fait;
-
-  Task(String nom) {
-    this.nom = nom;
-    this.fait = false;
   }
 }
